# apps/api/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/workspaces/{workspaceId}/projects/{projectId}/matrix": {
      "get": {
        "operationId": "ProjectMatrixController_getProjectMatrix",
        "parameters": [
          {
            "name": "x-task-user-id",
            "in": "header",
            "description": "Temporary trusted user context header until AuthModule owns request identity. Not an authentication mechanism.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectMatrixDto"
                }
              }
            }
          },
          "404": {
            "description": "Workspace or active project is missing or not visible."
          }
        },
        "summary": "Get the task matrix for a visible project",
        "tags": ["projects"]
      }
    },
...
    "/internal/telegram/context/resolve": {
      "post": {
        "operationId": "TelegramController_resolveContext",
        "parameters": [
          {
            "name": "x-task-bot-secret",
            "in": "header",
            "description": "Internal Telegram bot shared secret. Not a user authentication mechanism.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveTelegramContextDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelegramContextResolutionDto"
                }
              }
            }
          },
          "400": {
            "description": "Telegram context payload is invalid."
          },
          "401": {
            "description": "Telegram bot shared secret is missing or invalid."
          }
        },
        "summary": "Resolve Telegram bot message identity and chat context",
        "tags": ["telegram"]
      }
    },
    "/internal/telegram/confirmations/callback": {
      "post": {
        "operationId": "TelegramController_handleConfirmationCallback",
        "parameters": [
          {
            "name": "x-task-bot-secret",
            "in": "header",
            "description": "Internal Telegram bot shared secret. Not a user authentication mechanism.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelegramConfirmationCallbackDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelegramConfirmationCallbackResultDto"
                }
              }
            }
          },
          "400": {
            "description": "Telegram confirmation callback payload is invalid."
          },
          "401": {
            "description": "Telegram bot shared secret is missing or invalid."
          },
          "403": {
            "description": "Telegram user or chat is not linked to a workspace visible to the user."
          },
          "404": {
            "description": "Confirmation request is missing or not visible."
          }
        },
        "summary": "Apply a Telegram confirmation callback action",
        "tags": ["telegram"]
      }
    },
    "/telegram/mini-app/identity/link-status": {
      "get": {
        "operationId": "TelegramMiniAppController_getIdentityLinkStatus",
        "parameters": [
          {
            "name": "x-task-user-id",
            "in": "header",
            "description": "Temporary trusted user context header until AuthModule owns request identity. Not an authentication mechanism.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelegramIdentityLinkStatusDto"
                }
              }
            }
          },
          "403": {
            "description": "Current user has no linked Telegram identity."
          }
        },
        "summary": "Get the current user's Telegram Mini App identity link status",
        "tags": ["telegram"]
      }
    },
    "/telegram/mini-app/init-data/verify": {
      "post": {
        "operationId": "TelegramMiniAppController_verifyInitData",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyTelegramMiniAppInitDataDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifiedTelegramMiniAppInitDataDto"
                }
              }
            }
          },
          "400": {
            "description": "Telegram Mini App initData payload is malformed."
          },
          "401": {
            "description": "Telegram Mini App initData is invalid or expired."
          }
        },
        "summary": "Verify Telegram Mini App initData and return the stable identity",
        "tags": ["telegram"]
      }
    },
    "/telegram/mini-app/identity/link": {
      "post": {
        "operationId": "TelegramMiniAppController_linkIdentity",
        "parameters": [
          {
            "name": "x-task-user-id",
            "in": "header",
            "description": "Temporary trusted user context header until AuthModule owns request identity. Not an authentication mechanism.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyTelegramMiniAppInitDataDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkedTelegramIdentityDto"
                }
              }
            }
          },
          "400": {
            "description": "Telegram Mini App initData payload is malformed."
          },
          "401": {
            "description": "Telegram Mini App initData is invalid or expired."
          },
          "404": {
            "description": "Current user is missing."
          },
          "409": {
            "description": "Telegram identity is already linked to another user."
          }
        },
        "summary": "Link verified Telegram Mini App identity to the current user",
        "tags": ["telegram"]
      }
    },
...
    "/workspaces/{workspaceId}/agent/runs": {
      "get": {
        "operationId": "AgentRunsController_listWorkspaceRuns",
        "parameters": [
          {
            "name": "x-task-user-id",
            "in": "header",
            "description": "Temporary trusted user context header until AuthModule owns request identity. Not an authentication mechanism.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AgentRunSummaryDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Workspace is missing or not visible to the current user."
          }
        },
        "summary": "List current user agent run history in a visible workspace",
        "tags": ["agent"]
      }
    },
    "/workspaces/{workspaceId}/agent/runs/{agentRunId}": {
      "get": {
        "operationId": "AgentRunsController_getWorkspaceRun",
        "parameters": [
          {
            "name": "x-task-user-id",
            "in": "header",
            "description": "Temporary trusted user context header until AuthModule owns request identity. Not an authentication mechanism.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "agentRunId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunDetailDto"
                }
              }
            }
          },
          "404": {
            "description": "Workspace or agent run is missing or not visible to the current user."
          }
        },
        "summary": "Get one current user agent run with safe audit records",
        "tags": ["agent"]
      }
    }
  },
  "info": {
    "title": "tAsk API",
    "description": "Backend API contract for tAsk.",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "TaskSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "workspaceId": {
            "type": "string",
            "format": "uuid"
          },
          "projectId": {
            "type": "string",
            "format": "uuid"
          },
          "parentTaskId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "title": {
            "type": "string",
            "example": "Record bass"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "statusId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "assigneeUserId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "createdByUserId": {
            "type": "string",
            "format": "uuid"
          },
          "position": {
            "type": "string",
            "example": "1000"
          },
          "dueAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "sourceSkillId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "sourceSkillVersionId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          },
          "archivedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "workspaceId",
          "projectId",
          "title",
          "createdByUserId",
          "position",
          "metadata",
          "createdAt",
          "updatedAt"
        ]
      },
      "ProjectMatrixStageDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "name": {
            "type": "string",
            "example": "In progress"
          },
          "color": {
            "type": "string",
            "example": "#2563EB",
            "nullable": true
          },
          "position": {
            "type": "string",
            "example": "1000"
          },
          "isDone": {
            "type": "boolean"
          }
        },
        "required": ["name", "position", "isDone"]
      },
      "ProjectMatrixCellDto": {
        "type": "object",
        "properties": {
          "columnTaskId": {
            "type": "string",
            "format": "uuid"
          },
          "stageId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "tasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskSummaryDto"
            }
          }
        },
        "required": ["columnTaskId", "tasks"]
      },
      "ProjectMatrixDto": {
        "type": "object",
        "properties": {
          "columns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskSummaryDto"
            }
          },
          "stages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectMatrixStageDto"
            }
          },
          "cells": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectMatrixCellDto"
            }
          }
        },
        "required": ["columns", "stages", "cells"]
      },
...
      "ResolveTelegramContextDto": {
        "type": "object",
        "properties": {
          "telegramId": {
            "type": "string",
            "example": "123456789"
          },
          "telegramChatId": {
            "type": "string",
            "example": "-100987654321"
          }
        },
        "required": ["telegramId", "telegramChatId"]
      },
      "TelegramContextResolutionDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "resolved",
              "telegram_user_unlinked",
              "telegram_chat_unlinked",
              "user_not_in_chat_workspace"
            ]
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "workspaceId": {
            "type": "string",
            "format": "uuid"
          },
          "defaultProjectId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "required": ["status"]
      },
      "TelegramConfirmationCallbackDto": {
        "type": "object",
        "properties": {
          "telegramId": {
            "type": "string",
            "example": "123456789"
          },
          "telegramChatId": {
            "type": "string",
            "example": "-100987654321"
          },
          "confirmationRequestId": {
            "type": "string",
            "format": "uuid"
          },
          "action": {
            "type": "string",
            "enum": ["confirm", "cancel"]
          }
        },
        "required": ["telegramId", "telegramChatId", "confirmationRequestId", "action"]
      },
      "TelegramConfirmationCallbackResultDto": {
        "type": "object",
        "properties": {
          "confirmationRequestId": {
            "type": "string",
            "format": "uuid"
          },
          "action": {
            "type": "string",
            "enum": ["confirm", "cancel"]
          },
          "status": {
            "type": "string",
            "enum": ["confirmed", "cancelled"]
          }
        },
        "required": ["confirmationRequestId", "action", "status"]
      },
      "TelegramIdentityLinkStatusDto": {
        "type": "object",
        "properties": {
          "telegramId": {
            "type": "string",
            "example": "123456789"
          },
          "linkedAt": {
            "format": "date-time",
            "type": "string"
          },
          "lastSeenAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": ["telegramId", "linkedAt"]
      },
      "VerifyTelegramMiniAppInitDataDto": {
        "type": "object",
        "properties": {
          "initData": {
            "type": "string",
            "description": "Raw Telegram.WebApp.initData query string.",
            "maxLength": 8192
          }
        },
        "required": ["initData"]
      },
      "VerifiedTelegramMiniAppInitDataDto": {
        "type": "object",
        "properties": {
          "telegramId": {
            "type": "string",
            "example": "123456789"
          },
          "authDate": {
            "type": "string",
            "example": "1720468800"
          }
        },
        "required": ["telegramId", "authDate"]
      },
      "LinkedTelegramIdentityDto": {
        "type": "object",
        "properties": {
          "telegramId": {
            "type": "string",
            "example": "123456789"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": ["telegramId", "userId"]
      },
...
      "AgentRunSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "workspaceId": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "source": {
            "type": "string",
            "enum": ["telegram", "web", "mini_app"]
          },
          "sourceMessageId": {
            "type": "string",
            "nullable": true
          },
          "model": {
            "type": "string",
            "nullable": true
          },
          "inputText": {
            "type": "string",
            "example": "@task what is next for the album?"
          },
          "finalResponse": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": ["running", "waiting_confirmation", "completed", "failed"]
          },
          "error": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "workspaceId",
          "userId",
          "source",
          "inputText",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "AgentRunToolCallAuditDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "toolName": {
            "type": "string"
          },
          "arguments": {
            "type": "object",
            "additionalProperties": true
          },
          "result": {
            "type": "object",
            "additionalProperties": true,
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": ["pending", "success", "error"]
          },
          "error": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": ["id", "toolName", "arguments", "status", "createdAt"]
      },
      "AgentRunConfirmationLinkDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "type": "string"
          },
          "preview": {
            "type": "object",
            "additionalProperties": true
          },
          "status": {
            "type": "string",
            "enum": ["pending", "confirmed", "cancelled", "expired"]
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["id", "kind", "preview", "status", "expiresAt", "createdAt", "updatedAt"]
      },
      "AgentRunDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "workspaceId": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "source": {
            "type": "string",
            "enum": ["telegram", "web", "mini_app"]
          },
          "sourceMessageId": {
            "type": "string",
            "nullable": true
          },
          "model": {
            "type": "string",
            "nullable": true
          },
          "inputText": {
            "type": "string",
            "example": "@task what is next for the album?"
          },
          "finalResponse": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": ["running", "waiting_confirmation", "completed", "failed"]
          },
          "error": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "toolCalls": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentRunToolCallAuditDto"
            }
          },
          "confirmationRequests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentRunConfirmationLinkDto"
            }
          }
        },
        "required": [
          "id",
          "workspaceId",
          "userId",
          "source",
          "inputText",
          "status",
          "createdAt",
          "updatedAt",
          "toolCalls",
          "confirmationRequests"
        ]
      },
...
    }
  }
}